+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+- **CUMULUS-1455**
+  - CMR token links updated to point to CMR legacy services rather than echo
+
 - **CUMULUS-1211**
   - Errors thrown during granule discovery are no longer swallowed and ignored.
     Rather, errors are propagated to allow for proper error-handling and